app/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
fake-useragent==0.1.11
Flask==2.0.3
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
requests==2.27.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.0.3
